+++ src/realm/sync/noinst/client_impl_base.hpp
@@ -564,8 +564,9 @@ private:
     void initiate_disconnect_wait();
     void handle_disconnect_wait(Status status);
     void read_or_write_error(std::error_code ec, std::string_view msg);
-    void close_due_to_protocol_error(std::error_code, std::optional<std::string_view> msg = std::nullopt);
+    void close_due_to_protocol_error(Status status);
 
+    void close_due_to_transient_error(Status status, ConnectionTerminationReason reason);
     void close_due_to_client_side_error(Status status, IsFatal is_fatal, ConnectionTerminationReason reason);
     void close_due_to_client_side_error(std::error_code, std::optional<std::string_view> msg, IsFatal is_fatal,
                                         ConnectionTerminationReason reason);
@@ -584,7 +585,7 @@ private:
     void receive_mark_message(session_ident_type, request_ident_type);
     void receive_unbound_message(session_ident_type);
     void receive_test_command_response(session_ident_type, request_ident_type, std::string_view body);
-    void handle_protocol_error(ClientProtocol::Error);
+    void handle_protocol_error(ClientProtocol::Error, std::string message);
 
     // These are only called from Session class.
     void enlist_to_send(Session*);
@@ -1211,21 +1212,21 @@ private:
     void send_query_change_message();
     void send_json_error_message();
     void send_test_command_message();
-    std::error_code receive_ident_message(SaltedFileIdent);
-    void receive_download_message(const SyncProgress&, std::uint_fast64_t downloadable_bytes,
-                                  DownloadBatchState last_in_batch, int64_t query_version, const ReceivedChangesets&);
-    std::error_code receive_mark_message(request_ident_type);
-    std::error_code receive_unbound_message();
-    std::error_code receive_error_message(const ProtocolErrorInfo& info);
-    std::error_code receive_query_error_message(int error_code, std::string_view message, int64_t query_version);
-    std::error_code receive_test_command_response(request_ident_type, std::string_view body);
+    Status receive_ident_message(SaltedFileIdent);
+    Status receive_download_message(const SyncProgress&, std::uint_fast64_t downloadable_bytes,
+                                    DownloadBatchState last_in_batch, int64_t query_version,
+                                    const ReceivedChangesets&);
+    Status receive_mark_message(request_ident_type);
+    Status receive_unbound_message();
+    Status receive_error_message(const ProtocolErrorInfo& info);
+    Status receive_query_error_message(int error_code, std::string_view message, int64_t query_version);
+    Status receive_test_command_response(request_ident_type, std::string_view body);
 
     void initiate_rebind();
     void reset_protocol_state() noexcept;
     void ensure_enlisted_to_send();
     void enlist_to_send();
-    bool check_received_sync_progress(const SyncProgress&) noexcept;
-    bool check_received_sync_progress(const SyncProgress&, int&) noexcept;
+    Status check_received_sync_progress(const SyncProgress&) noexcept;
     void check_for_upload_completion();
     void check_for_download_completion();
 
@@ -1312,7 +1313,10 @@ inline void ClientImpl::Connection::voluntary_disconnect()
 {
     m_reconnect_info.update(ConnectionTerminationReason::closed_voluntarily, std::nullopt);
     constexpr bool try_again = true;
-    disconnect(SessionErrorInfo{Status{ClientError::connection_closed, "Connection closed"}, try_again}); // Throws
+    SessionErrorInfo error_info{Status{ErrorCodes::ConnectionClosed, "Connection closed"}, try_again};
+    error_info.server_requests_action = ProtocolErrorInfo::Action::Transient;
+
+    disconnect(std::move(error_info)); // Throws
 }
 
 inline void ClientImpl::Connection::involuntary_disconnect(const SessionErrorInfo& info,
@@ -1630,12 +1634,6 @@ inline void ClientImpl::Session::enlist_to_send()
     REALM_ASSERT(!m_enlisted_to_send);
     m_enlisted_to_send = true;
     m_conn.enlist_to_send(this); // Throws
-}
-
-inline bool ClientImpl::Session::check_received_sync_progress(const SyncProgress& progress) noexcept
-{
-    int error_code = 0; // Dummy
-    return check_received_sync_progress(progress, error_code);
 }
 
 } // namespace sync