+++ test/object-store/c_api/c_api.cpp
@@ -679,19 +679,19 @@ TEST_CASE("C API (non-database)", "[c_api]") {
         using namespace realm::sync;
         std::string message;
 
-        std::error_code error_code = make_error_code(sync::ClientError::connection_closed);
-        realm_sync_error_code_t error = c_api::to_capi(SystemError(error_code, "").to_status(), message);
+        Status fake_status(ErrorCodes::SyncProtocolInvariantFailed, "fake sync error");
+        realm_sync_error_code_t error = c_api::to_capi(fake_status, message);
         CHECK(error.category == realm_sync_error_category_e::RLM_SYNC_ERROR_CATEGORY_CLIENT);
-        CHECK(error.value == int(error_code.value()));
-        CHECK(error_code.message() == error.message);
+        CHECK(error.value == int(ErrorCodes::SyncProtocolInvariantFailed));
+        CHECK(fake_status.reason() == error.message);
         CHECK(message == error.message);
 
         std::error_code ec_check;
         c_api::sync_error_to_error_code(error, &ec_check);
-        CHECK(ec_check.category() == realm::sync::client_error_category());
-        CHECK(ec_check.value() == int(error_code.value()));
+        CHECK(ec_check.category() == std::generic_category());
+        CHECK(ec_check.value() == int(fake_status.code()));
 
-        error_code = make_error_code(sync::ProtocolError::connection_closed);
+        std::error_code error_code = make_error_code(sync::ProtocolError::connection_closed);
         error = c_api::to_capi(SystemError(error_code, "").to_status(), message);
         CHECK(error.category == realm_sync_error_category_e::RLM_SYNC_ERROR_CATEGORY_CONNECTION);